Refactored helper unit test classes
Extracted classes which provides data and verify test results into seperate files.

Replaced String with string and removed the System namespace.
